Reduce code duplication in LayoutNameSupport
The `templateLayoutName` preference is the same between Swing and SWT,
so the preference constants can be moved to the Core plugin. Due to
split packages, this then means that the `IPreferenceConstants` in the
Java plugin has to be removed, as it conflicts with the class in the
Core plugin. But which doesn't matter, as they are otherwise identical.

This then means that the `getTemplate()` can be moved up to the parent
class. Only the list of available templates needs to be known, which is
declared via the new, abstract `getTemplates()` method.

Additionally, the `getValueMap()` method can be partially moved to the
base class, because the `layoutAcronym` and `layoutClassName` styles are
shared. The method is overridden to add Swing and SWT-specific layouts.
