+++ org.eclipse.wb.core.java/META-INF/MANIFEST.MF
@@ -12,7 +12,7 @@ Require-Bundle: org.eclipse.ui;bundle-version="[3.206.0,4.0.0)",
  org.eclipse.debug.ui;bundle-version="[3.18.400,4.0.0)",
  org.eclipse.core.resources;bundle-version="[3.20.200,4.0.0)",
  org.eclipse.ui.editors;bundle-version="[3.17.300,4.0.0)",
- org.eclipse.wb.core;bundle-version="[1.22.0,2.0.0)",
+ org.eclipse.wb.core;bundle-version="[1.24.0,2.0.0)",
  org.eclipse.wb.core.databinding.xsd;bundle-version="[1.0.300,2.0.0)",
  org.eclipse.compare;bundle-version="[3.11.0,4.0.0)",
  org.eclipse.jdt.core;bundle-version="[3.38.0,4.0.0)";visibility:=reexport,
