+++ org.eclipse.wb.core/META-INF/MANIFEST.MF
@@ -342,7 +342,8 @@ Export-Package: org.eclipse.wb.core.branding,
    org.eclipse.wb.tests,
    org.eclipse.wb.swing,
    org.eclipse.wb.swt,
-   org.eclipse.wb.rcp",
+   org.eclipse.wb.rcp,
+   org.eclipse.wb.core.java",
  org.eclipse.wb.internal.core.preferences.bind;
   x-friends:="org.eclipse.wb.swing.FormLayout,
    org.eclipse.wb.swing.MigLayout,
