+++ org.eclipse.wb.swing/META-INF/MANIFEST.MF
@@ -106,9 +106,9 @@ Require-Bundle: org.eclipse.ui;bundle-version="[3.206.0,4.0.0)",
  org.eclipse.core.databinding.observable;bundle-version="[1.13.300,2.0.0)",
  org.eclipse.core.runtime;bundle-version="[3.31.100,4.0.0)",
  org.eclipse.core.resources;bundle-version="[3.20.200,4.0.0)",
- org.eclipse.wb.core;bundle-version="[1.21.0,2.0.0)";visibility:=reexport,
+ org.eclipse.wb.core;bundle-version="[1.24.0,2.0.0)";visibility:=reexport,
  org.eclipse.wb.core.ui;bundle-version="[1.10.800,2.0.0)";visibility:=reexport,
- org.eclipse.wb.core.java;bundle-version="[1.13.100,2.0.0)";visibility:=reexport,
+ org.eclipse.wb.core.java;bundle-version="[1.15.0,2.0.0)";visibility:=reexport,
  org.eclipse.draw2d;bundle-version="[3.20.0,4.0.0)",
  org.eclipse.gef;bundle-version="[3.21.100,4.0.0)",
  org.eclipse.nebula.widgets.cdatetime;bundle-version="[1.5.0,2.0.0)"
