+++ org.eclipse.wb.swing/src/org/eclipse/wb/internal/swing/model/layout/LayoutNameSupport.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011 Google, Inc.
+ * Copyright (c) 2011, 2026 Google, Inc. and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -13,12 +13,8 @@
 package org.eclipse.wb.internal.swing.model.layout;
 
 import org.eclipse.wb.internal.swing.model.component.ContainerInfo;
-import org.eclipse.wb.internal.swing.preferences.IPreferenceConstants;
-
-import org.eclipse.jface.preference.IPreferenceStore;
 
 import java.util.Map;
-import java.util.TreeMap;
 
 /**
  * Support for managing name of {@link LayoutInfo}, so that it corresponds to the name of its parent
@@ -27,14 +23,12 @@ import java.util.TreeMap;
  * @author sablin_aa
  * @coverage swing.model.layout
  */
-public final class LayoutNameSupport
-extends
-org.eclipse.wb.internal.core.model.layout.LayoutNameSupport<LayoutInfo> {
-	public final static String[] TEMPLATES = new String[]{
-			"${layoutAcronym}_${containerName}",
-			"${layoutAcronym}${containerName-cap}",
-			"${containerName}${layoutClassName}",
-	"${defaultName}"};
+public final class LayoutNameSupport extends org.eclipse.wb.internal.core.model.layout.LayoutNameSupport<LayoutInfo> {
+	public final static String[] TEMPLATES = new String[] { //
+			"${layoutAcronym}_${containerName}", //
+			"${layoutAcronym}${containerName-cap}", //
+			"${containerName}${layoutClassName}", //
+			"${defaultName}" };
 
 	////////////////////////////////////////////////////////////////////////////
 	//
@@ -51,22 +45,15 @@ org.eclipse.wb.internal.core.model.layout.LayoutNameSupport<LayoutInfo> {
 	//
 	////////////////////////////////////////////////////////////////////////////
 	@Override
-	protected String getTemplate() {
-		IPreferenceStore preferences = m_childInfo.getDescription().getToolkit().getPreferences();
-		String template = preferences.getString(IPreferenceConstants.P_LAYOUT_NAME_TEMPLATE);
-		if (!isValidTemplate(TEMPLATES, template)) {
-			template = getTemplateForDefault();
-		}
-		return template;
+	protected String[] getTemplates() {
+		return TEMPLATES;
 	}
 
 	@Override
 	protected Map<String, String> getValueMap() {
 		// prepare variables
-		Map<String, String> valueMap = new TreeMap<>();
+		Map<String, String> valueMap = super.getValueMap();
 		{
-			valueMap.put("layoutAcronym", getAcronym());
-			valueMap.put("layoutClassName", getClassName());
 			valueMap.put("containerName", getParentName());
 			valueMap.put("containerName-cap", getParentNameCap());
 		}
