+++ org.eclipse.wb.swing/src/org/eclipse/wb/internal/swing/preferences/IPreferenceConstants.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011 Google, Inc.
+ * Copyright (c) 2011, 2026 Google, Inc. and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -18,12 +18,6 @@ package org.eclipse.wb.internal.swing.preferences;
  * @author scheglov_ke
  * @coverage swing.preferences
  */
-public interface IPreferenceConstants
-extends
-org.eclipse.wb.internal.core.preferences.IPreferenceConstants {
+public interface IPreferenceConstants extends org.eclipse.wb.internal.core.preferences.IPreferenceConstants {
 	String TOOLKIT_ID = "org.eclipse.wb.swing";
-	/**
-	 * The template for <code>Layout</code> variable name.
-	 */
-	String P_LAYOUT_NAME_TEMPLATE = "templateLayoutName";
 }
