+++ org.eclipse.wb.swt/META-INF/MANIFEST.MF
@@ -54,14 +54,14 @@ Export-Package: org.eclipse.wb.internal.swt;x-friends:="org.eclipse.wb.os.linux,
  org.eclipse.wb.internal.swt.model.widgets.menu;x-friends:="org.eclipse.wb.tests",
  org.eclipse.wb.internal.swt.palette;x-friends:="org.eclipse.wb.tests",
  org.eclipse.wb.internal.swt.parser,
- org.eclipse.wb.internal.swt.preferences,
+ org.eclipse.wb.internal.swt.preferences;x-friends:="org.eclipse.wb.tests,org.eclipse.wb.rcp",
  org.eclipse.wb.internal.swt.preferences.layout,
  org.eclipse.wb.internal.swt.support,
  org.eclipse.wb.internal.swt.utils
 Require-Bundle: org.eclipse.ui;bundle-version="[3.206.0,4.0.0)",
  org.eclipse.core.runtime;bundle-version="[3.31.100,4.0.0)",
- org.eclipse.wb.core;bundle-version="[1.22.0,2.0.0)",
- org.eclipse.wb.core.java;bundle-version="[1.13.100,2.0.0)",
+ org.eclipse.wb.core;bundle-version="[1.24.0,2.0.0)",
+ org.eclipse.wb.core.java;bundle-version="[1.15.0,2.0.0)",
  org.eclipse.pde.core;bundle-version="[3.18.100,4.0.0)",
  org.eclipse.draw2d;bundle-version="[3.20.0,4.0.0)",
  org.eclipse.gef;bundle-version="[3.21.100,4.0.0)",
