+++ org.eclipse.wb.swt/src/org/eclipse/wb/internal/swt/model/layout/LayoutNameSupport.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011 Google, Inc.
+ * Copyright (c) 2011, 2026 Google, Inc. and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -13,12 +13,7 @@
 package org.eclipse.wb.internal.swt.model.layout;
 
 import org.eclipse.wb.internal.swt.model.widgets.CompositeInfo;
-import org.eclipse.wb.internal.swt.preferences.IPreferenceConstants;
-
-import org.eclipse.jface.preference.IPreferenceStore;
-
 import java.util.Map;
-import java.util.TreeMap;
 
 /**
  * Support for managing name of {@link LayoutInfo}, so that it corresponds to the name of its parent
@@ -27,14 +22,12 @@ import java.util.TreeMap;
  * @author sablin_aa
  * @coverage swt.model.layout
  */
-public final class LayoutNameSupport
-extends
-org.eclipse.wb.internal.core.model.layout.LayoutNameSupport<LayoutInfo> {
-	public final static String[] TEMPLATES = new String[]{
-			"${layoutAcronym}_${compositeName}",
-			"${layoutAcronym}${compositeName-cap}",
-			"${compositeName}${layoutClassName}",
-	"${defaultName}"};
+public final class LayoutNameSupport extends org.eclipse.wb.internal.core.model.layout.LayoutNameSupport<LayoutInfo> {
+	public final static String[] TEMPLATES = new String[] { //
+			"${layoutAcronym}_${compositeName}", //
+			"${layoutAcronym}${compositeName-cap}", //
+			"${compositeName}${layoutClassName}", //
+			"${defaultName}" };
 
 	////////////////////////////////////////////////////////////////////////////
 	//
@@ -51,22 +44,15 @@ org.eclipse.wb.internal.core.model.layout.LayoutNameSupport<LayoutInfo> {
 	//
 	////////////////////////////////////////////////////////////////////////////
 	@Override
-	protected String getTemplate() {
-		IPreferenceStore preferences = m_childInfo.getDescription().getToolkit().getPreferences();
-		String template = preferences.getString(IPreferenceConstants.P_LAYOUT_NAME_TEMPLATE);
-		if (!isValidTemplate(TEMPLATES, template)) {
-			template = getTemplateForDefault();
-		}
-		return template;
+	protected String[] getTemplates() {
+		return TEMPLATES;
 	}
 
 	@Override
 	protected Map<String, String> getValueMap() {
 		// prepare variables
-		Map<String, String> valueMap = new TreeMap<>();
+		Map<String, String> valueMap = super.getValueMap();
 		{
-			valueMap.put("layoutAcronym", getAcronym());
-			valueMap.put("layoutClassName", getClassName());
 			valueMap.put("compositeName", getParentName());
 			valueMap.put("compositeName-cap", getParentNameCap());
 		}
