+++ org.eclipse.wb.swt/src/org/eclipse/wb/internal/swt/preferences/IPreferenceConstants.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011, 2023 Google, Inc.
+ * Copyright (c) 2011, 2026 Google, Inc. and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -13,7 +13,6 @@
 package org.eclipse.wb.internal.swt.preferences;
 
 import org.eclipse.wb.internal.swt.model.layout.LayoutDataInfo;
-import org.eclipse.wb.internal.swt.model.layout.LayoutInfo;
 
 import org.eclipse.jface.resource.LocalResourceManager;
 
@@ -24,7 +23,7 @@ import org.eclipse.jface.resource.LocalResourceManager;
  * @author lobas_av
  * @coverage swt.preferences
  */
-public interface IPreferenceConstants {
+public interface IPreferenceConstants extends org.eclipse.wb.internal.core.preferences.IPreferenceConstants {
 	////////////////////////////////////////////////////////////////////////////
 	//
 	// Preferences
@@ -35,10 +34,6 @@ public interface IPreferenceConstants {
 	 * use it for color/font/image access. This allows use resources sharing.
 	 */
 	String P_USE_RESOURCE_MANAGER = "useResourceManager";
-	/**
-	 * The template for {@link LayoutInfo} name.
-	 */
-	String P_LAYOUT_NAME_TEMPLATE = "templateLayoutName";
 	/**
 	 * The template for {@link LayoutDataInfo} name.
 	 */
